# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API con Express y TypeScript",
    "version": "1.0.0",
    "description": "Documentación generada con swagger-jsdoc"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/direccion": {
      "get": {
        "summary": "Obtiene todas las direcciones",
        "tags": [
          "Direccion"
        ],
        "responses": {
          "200": {
            "description": "Lista de direcciones"
          }
        }
      },
      "post": {
        "summary": "Crea una dirección nueva (requiere autenticación)",
        "tags": [
          "Direccion"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Datos para crear dirección",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "calle",
                  "numero",
                  "codigo_postal",
                  "user_id",
                  "localidad_id"
                ],
                "properties": {
                  "codigo_postal": {
                    "type": "number"
                  },
                  "numero": {
                    "type": "number"
                  },
                  "calle": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "number"
                  },
                  "localidad_id": {
                    "type": "number"
                  }
                },
                "example": {
                  "codigo_postal": 5515,
                  "numero": 290,
                  "calle": "Capetillo",
                  "user_id": 2,
                  "localidad_id": 4,
                  "activo": true
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dirección creada"
          },
          "400": {
            "description": "Error en los datos enviados"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/direccion/{id}": {
      "get": {
        "summary": "Obtiene una dirección por ID",
        "tags": [
          "Direccion"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la dirección"
          }
        ],
        "responses": {
          "200": {
            "description": "Dirección encontrada"
          },
          "404": {
            "description": "Dirección no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualiza una dirección por ID (requiere autenticación)",
        "tags": [
          "Direccion"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la dirección"
          }
        ],
        "requestBody": {
          "description": "Datos para actualizar dirección",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "codigo_postal": {
                    "type": "number"
                  },
                  "numero": {
                    "type": "number"
                  },
                  "calle": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "number"
                  },
                  "localidad_id": {
                    "type": "number"
                  },
                  "activo": {
                    "type": "boolean"
                  }
                },
                "example": {
                  "codigo_postal": 5500,
                  "numero": 123,
                  "calle": "Ozamis",
                  "user_id": 1,
                  "localidad_id": 5
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dirección actualizada"
          },
          "400": {
            "description": "Error en los datos enviados"
          },
          "404": {
            "description": "No se encontró la dirección"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      },
      "patch": {
        "summary": "Activa o desactiva una dirección (requiere autenticación)",
        "tags": [
          "Direccion"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la dirección"
          }
        ],
        "requestBody": {
          "description": "Estado activo o inactivo",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activo": {
                    "type": "boolean"
                  }
                },
                "example": {
                  "activo": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "400": {
            "description": "No se pudo patchear la direccion"
          },
          "404": {
            "description": "No se encontro la direccion"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
...
    "/imagenes": {
      "get": {
        "summary": "Obtiene todas las imágenes",
        "tags": [
          "Imágenes"
        ],
        "responses": {
          "200": {
            "description": "Lista de imágenes"
          }
        }
      },
      "post": {
        "summary": "Crea una nueva imagen (requiere autenticación)",
        "tags": [
          "Imágenes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string"
                  },
                  "activo": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "url"
                ]
              },
              "example": {
                "url": "https://assets.adidas.com/images/w_600,f_auto,q_auto/5002fee0521a4b87b878f79f066afb34_9366/Zapatillas_Handball_Spezial_Granate_JP8726_00_plp_standard.jpg"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Imagen creada"
          },
          "500": {
            "description": "Error al crear la imagen"
          }
        }
      }
    },
    "/imagenes/{id}": {
      "get": {
        "summary": "Obtiene una imagen por ID",
        "tags": [
          "Imágenes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID de la imagen",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Imagen encontrada"
          },
          "404": {
            "description": "Imagen no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualiza una imagen por ID (requiere autenticación)",
        "tags": [
          "Imágenes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID de la imagen",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Datos para actualizar la imagen",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string"
                  },
                  "activo": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "url": "https://assets.adidas.com/images/w_600,f_auto,q_auto/2fb23ba467e54b1fb3469c7824bbc40e_9366/Zapatillas_adidas_Grand_Court_Base_00s_Negro_IH6184_00_plp_standard.jpg"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen actualizada"
          },
          "404": {
            "description": "Imagen no encontrada"
          },
          "500": {
            "description": "Error al actualizar la imagen"
          }
        }
      },
      "patch": {
        "summary": "Activa o desactiva una imagen (requiere autenticación)",
        "tags": [
          "Imágenes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "ID de la imagen",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Estado activo o inactivo",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "activo": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          },
          "404": {
            "description": "Imagen no encontrada"
          },
          "500": {
            "description": "Error al patchear la imagen"
          }
        }
      }
    },
...
  },
...
}